Lecture 12 - Constants, Build Variants, Release Builds
