+++ app/build.gradle
@@ -20,8 +20,18 @@ android {
 
     buildTypes {
         release {
-            minifyEnabled false
+            minifyEnabled true
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+
+    flavorDimensions "datasource"
+    productFlavors {
+        firebase {
+            dimension "datasource"
+        }
+        twitter {
+            dimension "datasource"
         }
     }
 }
